Debug touch error and cleaned up packages. Pair Programming
