--- src/main/java/StringLiterals.java
@@ -1,32 +1,0 @@
-//TODO Javadocs
-public class StringLiterals { //TODO 1) categorize them 2) ABC by category
-  public final String ACCESS_LEVEL = (" ~ % ");
-  public final String BLANK_SCREEN = ("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n");
-
-  public final String COMPUTER_PROMPT = ("Please enter your Computer Name: ");
-  public final String EXIT =("Saving session...\n...copying shared history..\n...saving history...truncating history files...\n...completed.\n\n[Process completed]");
-
-  public final String INVALID_CD = ("cd: no such file or directory: ");
-
-  public final String USER_PROMPT = ("Please enter your User Name: ");
-
-  public final String WELCOME = ("Hello! Welcome to Linux Learner. If you are new to the Linux Command line then this is the place for you.\nIf you are not a Linux beginner, then let's test your knowledge with some tasks!");
-
-  public final String DIRECTORY_CONTENTS = ("Pictures Desktop Public Documents Downloads Library Movies");
-
-  public final String PWD = ("/Users/");
-
-  public final String LS_DESCRIPTION = ("List Directory Contents (ls):  List information about the FILEs (the current directory by default)./n Sort entries alphabetically if no order specified");
-
-  public final String EXIT_DESCRIPTION = ("Exit: Disconnects ssh user and Exits the terminal");
-  public final String HELP = ("help for that command is not available. You may use clear, ls, pwd, touch, or exit");
-
-  public final String PWD_DESCRIPTION = ("Print Working Directory (pwd): Print the full filename of the current working directory.");
-
-  public final String TOUCH_DESCRIPTION = ("Update the access and modification times of each FILE to the current time. /n A FILE argument that does not exist is created empty\n");//TODO clean these up
-
-  public final String CD_DESCRIPTION = ("Change Directory (cd): Change the current directory user is in to requested directory./n Note: user may not have permissions to access all Directories");
-
-  public final String CLEAR_DESCRIPTION = ("Clear: clears the screen, does not remove any data currently or previously used. ");
-
-}
