+++ src/main/java/com/linux/controller/DateTimeGroup.java
@@ -1,3 +1,5 @@
+package com.linux.controller;
+
 import java.text.SimpleDateFormat;
 import java.util.Date;
 //TODO javadocs
