+++ src/main/java/com/linux/view/Main.java
@@ -1,3 +1,7 @@
+package com.linux.view;
+
+import com.linux.controller.*;
+
 //TODO javadocs then maybe a test class
 public class Main {
