+++ src/main/java/com/linux/view/userTerminal.java
@@ -1,0 +1,5 @@
+package com.linux.view;
+
+public interface userTerminal {
+
+}
